move slack plugin to its own package
(cherry picked from commit 87611581121b551aaaf6ca765f0c557bd44ae92a)
