+++ plugins/slack/plugin_test.go
@@ -1,4 +1,4 @@
-package plugins
+package slack
 
 import (
 	"errors"
